Working pause/unpause!
But with a simple "i = i + 1" you can see weird behavior.
This is the threading stuff your mama warned you about.
You press "pause", it triggers between the i+1 and the assignment.
In the console, you examine i and see the old value 76. You set i
to 1000. Then you unpause, expecting i to be 1000. It immediately
does the store, of 77 into i. Doh.